+++ A4/Mesh.hpp
@@ -32,10 +32,10 @@ class Mesh : public Primitive
 {
 public:
     Mesh(const std::string &fname);
-    std::pair<glm::vec3, glm::vec3> *intersect(std::pair<glm::vec3, glm::vec3> ray) override;
+    std::pair<glm::dvec3, glm::dvec3> *intersect(std::pair<glm::dvec3, glm::dvec3> ray) override;
 
 private:
-    std::vector<glm::vec3> m_vertices;
+    std::vector<glm::dvec3> m_vertices;
     std::vector<Triangle> m_faces;
     NonhierBox bounding_box;
