+++ A4/PhongMaterial.hpp
@@ -9,14 +9,14 @@
 class PhongMaterial : public Material
 {
 public:
-    PhongMaterial(const glm::vec3 &kd, const glm::vec3 &ks, double shininess);
-    PhongMaterial(const glm::vec3 &kd, const glm::vec3 &ks, double shininess, float index_of_refraction, double transparency);
+    PhongMaterial(const glm::dvec3 &kd, const glm::dvec3 &ks, double shininess);
+    PhongMaterial(const glm::dvec3 &kd, const glm::dvec3 &ks, double shininess, double index_of_refraction, double transparency);
     virtual ~PhongMaterial();
 
-    glm::vec3 m_kd;
-    glm::vec3 m_ks;
+    glm::dvec3 m_kd;
+    glm::dvec3 m_ks;
 
     double m_shininess;
-    float m_index_of_refraction;
+    double m_index_of_refraction;
     double m_transparency;
 };
